refactor: restructure button components under 'action'
Moved 'button' folder to 'action' and updated imports across files to reflect the new structure. Also replaced all generic `useRef()` calls with `useRef(null)` for improved default handling.
